% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_source.R
\name{build_local}
\alias{build_local}
\title{Build source packages (vectorized over `pkgs`).}
\usage{
build_local(pkgs, src, ...)
}
\arguments{
\item{pkgs}{Character vector: Paths to each package to build.}

\item{src}{Length-1 string: Path to a directory to save the output.}

\item{...}{Arguments to [devtools::build()].}
}
\value{
First argument, invisible.
}
\description{
Build source packages (vectorized over `pkgs`).
}
\examples{
\dontrun{
build_local(fgeo_pkgs(), fgeo_src())
}
}
