% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_packages}
\alias{cran_packages}
\title{CRAN dependencies of \strong{fgeo} packages.}
\format{An object of class \code{character} of length 28.}
\usage{
cran_packages
}
\description{
CRAN dependencies of \strong{fgeo} packages.
}
\keyword{datasets}
