% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{dependencies_ls}
\title{Find all dependencies_ls of some packages.}
\usage{
dependencies(pkgs, exclude = NULL, section = "Imports")

dependencies_ls(pkgs, section = "Imports")
}
\arguments{
\item{pkgs}{Character vector: The name of installed packages.}

\item{exclude}{String. Pattern to exclude from output.}

\item{section}{Length-1 character vector: Section of DESCRITION to include.}
}
\value{
Character vector.
}
\description{
Find all dependencies_ls of some packages.
}
\examples{
pkgs <- fgeo::fgeo_dependencies("fgeo", FALSE)
dependencies(pkgs)
dependencies(pkgs, "Imports")
dependencies(pkgs, "Suggests")
dependencies(pkgs, exclude = "fgeo")
}
\author{
Adapted from \strong{tidyverse}.
}
