% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.install-package.R
\docType{package}
\name{fgeo.install-package}
\alias{fgeo.install}
\alias{fgeo.install-package}
\title{fgeo.install: A Helper For Installing fgeo}
\description{
A helper for installing fgeo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.install}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.install/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item Bier Ekaphan Kraichak \email{ekaphan.k@ku.th} [reviewer]
  \item Erika Gonzalez-Akre \email{GonzalezEB@si.edu@si.edu} [reviewer]
  \item Valentine Herrmann \email{HerrmannV@si.edu} [reviewer]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder]
}

}
\keyword{internal}
