% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo}
\alias{fgeo}
\title{fgeo packages.}
\usage{
fgeo(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. \code{FALSE} excludes "fgeo".}
}
\value{
Character vector.
}
\description{
Simpler version of \code{fgeo::fgeo()}.
}
\examples{
fgeo()
fgeo(FALSE)
}
