% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{fgeo_deps}
\alias{fgeo_deps}
\title{Shortcut to get all dependencies of fgeo packages, including fgeo itself.}
\usage{
fgeo_deps()
}
\value{
A character vector.
}
\description{
Shortcut to get all dependencies of fgeo packages, including fgeo itself.
}
\examples{
fgeo_deps()
}
