% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_install.R
\name{fgeo_install}
\alias{fgeo_install}
\title{Install fgeo from source code.}
\usage{
fgeo_install()
}
\value{
No return.
}
\description{
Install fgeo from source code.
}
\examples{
\dontrun{
fgeo_install()
}
}
