% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fgeo_packages}
\alias{fgeo_packages}
\title{Name of \strong{fgeo} packages.}
\format{An object of class \code{character} of length 9.}
\usage{
fgeo_packages
}
\description{
Name of \strong{fgeo} packages.
}
\keyword{datasets}
