% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{fgeo_source}
\alias{fgeo_source}
\title{Paths to source packages, in the order they should be installed.}
\usage{
fgeo_source()
}
\value{
Character vector.
}
\description{
Paths to source packages, in the order they should be installed.
}
\examples{
fgeo_source()
}
