% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_up.R
\name{follow_up}
\alias{follow_up}
\title{What you may do after installing fgeo.}
\description{
After installing \strong{fgeo} please visit its
\href{https://forestgeo.github.io/fgeo/}{website}. If you had any issue during
installation or please report it
\href{https://github.com/forestgeo/fgeo.install/issues/new}{here} or email me at
maurolepore@gmail.com.

\preformatted{# Warning: Updating packages may be slow if many packages are behind

# Check for uptates
update.packages()

# Similar but more comprehensive
remotes::update_packages()

# Best if you already had an old version of fgeo
remove.packages("fgeo")
fgeo.install::install_github2()
}
}
