% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_installation.R
\name{guide_installation}
\alias{guide_installation}
\title{Give installation instructions.}
\usage{
guide_installation(pkgs)
}
\arguments{
\item{pkgs}{Character vector: The names of the packages to install.}
}
\value{
Text output as with \code{cat()} and invisible \code{pkgs}.
}
\description{
Give installation instructions.
}
\examples{
guide_installation(c("pkg1", "pkg2"))
}
