% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install \strong{fgeo} from GitHub.}
\usage{
install(upgrade = "never", auth_token = .github_pat, ...)
}
\arguments{
\item{upgrade}{One of "ask", "always" or "never". "ask" prompts the user for
which out of date packages to upgrade. For non-interactive sessions "ask" is
equivalent to "always". \code{TRUE} and \code{FALSE} are also accepted and
correspond to "always" and "never" respectively.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/applications} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable.}

\item{...}{Other arguments passed to \link[remotes:install_github]{remotes::install_github}.}
}
\description{
This is a lazy version of \code{\link[remotes:install_github]{remotes::install_github()}}, customized to install
\strong{fgeo}. It is optimized for speed and
to increase the chances that the installation will succeed.
\itemize{
\item It informs what is the expected R environment.
\item It doesn't upgrade packages; instead it informs how you can do that.
}
}
\examples{
\dontrun{
install()
}
}
\seealso{
\link[remotes:install_github]{remotes::install_github}, \code{\link[=install_dependencies]{install_dependencies()}},
\code{\link[=install_fgeo]{install_fgeo()}}.
}
