% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{install_fgeo}
\alias{install_fgeo}
\title{Install \strong{fgeo} from GitHub.}
\usage{
install_fgeo(ref = "master")
}
\arguments{
\item{ref}{Desired git reference. Could be a commit, tag, a call to
github_pull(), or most commonly a branch name such as "master" or "dev".}
}
\description{
This function installs \strong{fgeo} dependencies from CRAN and fgeo packages from
GitHub. It installs only the packages that are not already installed.
}
\examples{
\dontrun{
install_fgeo()

# Install the development (maybe failing) branch of each package.
install_fgeo("dev")
}
}
\seealso{
\link[remotes:install_github]{remotes::install_github}, \link{follow_up}.
}
