% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_from_cache}
\alias{install_from_cache}
\alias{install_from_github}
\title{Install \strong{fgeo} from GitHub or cache.}
\usage{
install_from_cache()

install_from_github()
}
\description{
These functions install fgeo dependencies from CRAN and fgeo packages from
either GitHub or from a cache stored inside \strong{fgeo.install}. They install
only the packages that are not already installed.
}
\examples{
\dontrun{
install_from_github()
install_from_cache()
}
}
\seealso{
\link[remotes:install_github]{remotes::install_github}, \code{\link[remotes:update_packages]{remotes::update_packages()}},
\code{\link[utils:update.packages]{utils::update.packages()}}.
}
