% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github2.R
\name{install_github2}
\alias{install_github2}
\title{Install \strong{fgeo} from GitHub.}
\usage{
install_github2()
}
\description{
This function installs \strong{fgeo} dependencies from CRAN and fgeo packages from
either GitHub. It installs only the packages that are not already installed.
}
\examples{
\dontrun{
install_github2()
}
}
\seealso{
\link[remotes:install_github]{remotes::install_github}, \link{follow_up}.
}
