% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{path_source}
\alias{path_source}
\alias{fgeo_source}
\title{Path to source fgeo packages.}
\usage{
path_source(path = NULL)

fgeo_source()
}
\arguments{
\item{path}{Character vector: Path to a file in inst/extdata/source.}
}
\value{
Character vector.
}
\description{
Path to source fgeo packages.
}
\examples{
path_source()

path_source(fgeo.install::scheduled_packages)
# Evocative shortcut
fgeo_source()
}
