% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{prefix}
\alias{prefix}
\title{Prefix and separate matching patterns.}
\usage{
prefix(x, .prefix, sep = "_")
}
\arguments{
\item{x}{A character vector.}

\item{sep}{Separator to surround the pattern with.}

\item{prefix}{A character vector with the format
`c("prefix1" = "pattern1", "prefix2" = "pattern2")`.}
}
\value{
A character vector.
}
\description{
Takes effect only on elements that match the patterns.
}
\examples{
x <- letters[1:3]
prefix <- c(
 "01" = "c",
 "01" = "b",
 "02" = "a"
)

prefix(x, prefix)
prefix(x, prefix, sep = "")

# Useful to reorder
sort(x)
x2 <- prefix(x, prefix, sep = "")
sort(x2)

# Useful to plan the order in which these packages should be installed
pkgs <- fgeo::fgeo_dependencies("fgeo", include_self = FALSE)
plan <- c(
  "01" = "fgeo.x",
  "02" = "fgeo.base",
  "03" = "fgeo.ctfs",
  "04" = "fgeo.tool",
  "05" = "fgeo.map"
)
prefix(pkgs, plan)
}
