% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{Schedule fgeo packages in the order they should be installed.}
\usage{
schedule(pkgs, plan)
}
\arguments{
\item{pkgs}{}
}
\description{
Schedule fgeo packages in the order they should be installed.
}
\examples{
src <- fs::dir_ls("inst/extdata/source")
schedule(src, plan)

pkgs <- fgeo::fgeo_dependencies("fgeo", include_self = FALSE)
schedule(pkgs, plan)
}
