% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled_packages.R
\docType{data}
\name{scheduled_packages}
\alias{scheduled_packages}
\title{Source fgeo packages, ordered as they should be installed.}
\format{An object of class \code{character} of length 9.}
\usage{
scheduled_packages
}
\description{
Source fgeo packages, ordered as they should be installed.
}
\keyword{datasets}
