% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_source.R
\name{build_source}
\alias{build_source}
\title{Build source packages (vectorized over `pkgs`).}
\usage{
build_source(pkgs, src, branch = "master")
}
\arguments{
\item{pkgs}{Character vector: Paths to each package to build.}

\item{src}{Length-1 string: Path to a directory to save the output.}
}
\value{
First argument, invisible.
}
\description{
Build source packages (vectorized over `pkgs`).
}
\examples{
\dontrun{
build_source(fgeo_pkgs("../"), fgeo_src("../"))

withr::with_dir(
 "../fgeo",
  build_source(fgeo_pkgs(), fgeo_src())
)
}
}
