% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\docType{data}
\name{fgeo_packages}
\alias{fgeo_packages}
\alias{cran_packages}
\title{Installed packages.}
\format{An object of class \code{character} of length 9.}
\usage{
fgeo_packages

cran_packages
}
\description{
\itemize{
\item \code{fgeo_packages}: \strong{fgeo} packages.
\item \code{cran_packages}: Dependencies of \strong{fgeo} packages.
}
}
\examples{
fgeo_packages
cran_packages
}
\keyword{datasets}
