% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_installation.R
\name{guide_installation}
\alias{guide_installation}
\title{Give installation instructions.}
\usage{
guide_installation(pkgs,
  msg_issue = "You need to install the following packages from CRAN:",
  msg_hint = "Click 'Session > Restart R' to start a clean R session, then run:",
  bullets = cat_bullet(cry_warn(format(pkgs))))
}
\arguments{
\item{pkgs}{Character vector: The names of the packages to install.}

\item{msg_issue}{Character vector: A message about the general issue.}

\item{msg_hint}{Character vector: A message about the code to run.}

\item{bullets}{Bullets (see default).}
}
\value{
Text output as with \code{cat()} and invisible \code{pkgs}.
}
\description{
Give installation instructions.
}
\examples{
guide_installation(c("pkg1", "pkg2"))
}
