% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{install_fgeo}
\alias{install_fgeo}
\title{Ask users to install fgeo-dependencies from CRAN.}
\usage{
install_fgeo(pkgs = fgeo_source())
}
\arguments{
\item{pkgs}{Character vector: Path to source packages to install.}
}
\value{
Invisible \code{NULL}.
}
\description{
Ask users to install fgeo-dependencies from CRAN.
}
\examples{
\dontrun{
# Two packages from the top of the installation schedule.
install_fgeo(fgeo_source()[1:2])

# All packags
install_fgeo()
}
}
