% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{install_fgeo}
\alias{install_fgeo}
\title{Install fgeo from source code.}
\usage{
install_fgeo()
}
\value{
No return.
}
\description{
Install fgeo from source code.
}
\examples{
\dontrun{
install_fgeo()
}
}
