% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doing.R
\name{install_fgeo}
\alias{install_fgeo}
\title{Ask users to install fgeo-dependencies from CRAN.}
\usage{
install_fgeo(...)
}
\value{
Invisible \code{NULL}.
}
\description{
Ask users to install fgeo-dependencies from CRAN.
}
\examples{
\dontrun{
install_fgeo()
}
}
