% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fgeo.R
\name{install_fgeo}
\alias{install_fgeo}
\alias{install_dependencies}
\title{Install fgeo packages and ask or install their dependencies from CRAN.}
\usage{
install_fgeo(pkgs = fgeo_source())

install_dependencies(...)
}
\arguments{
\item{pkgs}{Character vector: Path to source packages to install.}

\item{...}{Arguments passed to \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\value{
\itemize{
\item \code{install_dependencies()} returns invisible \code{NULL}.
\item \code{install_fgeo()} returns invisible \code{pkgs}.
}
}
\description{
\itemize{
\item \code{install_dependencies()} installs from CRAN the dependencies of all
\strong{fgeo} packages.
\item \code{install_fgeo()} installs all \strong{fgeo} packages from source code stored in
\strong{fgeo.install}_. It requests all CRAN dependencies to be already
installed.
}
}
\examples{
\dontrun{
install_fgeo()
install_dependencies()
}
}
