% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{Sort packages in the order they should be installed.}
\usage{
schedule(pkgs, plan)
}
\arguments{
\item{pkgs}{Character vector: Normally the paths to source-packages to be
installed.}

\item{plan}{Character vector. See examples.}
}
\value{
Character vector ordered as packages should be installed.
}
\description{
Sort packages in the order they should be installed.
}
\examples{
plan <- c(
  "01" = "fgeo.x",
  "02" = "fgeo.base",
  "03" = "fgeo.ctfs",
  "04" = "fgeo.tool",
  "05" = "fgeo.map"
)

schedule(fgeo_source(), plan)

# WARNING: Bad order -- for demonstration
# The rescheduling affects only the planed pakcages. Then alpha ordered.
plan <- c(
  "01" = "fgeo.map",
  "02" = "fgeo.x"
)

schedule(fgeo_source(), plan)
}
