% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_fgeo_source}
\alias{update_fgeo_source}
\title{Build fgeo source-packages from github into a local directory.}
\usage{
update_fgeo_source()
}
\value{
Source packages (.tar.gz files).
}
\description{
Build fgeo source-packages from github into a local directory.
}
\examples{
\dontrun{
update_fgeo_source()
}
}
